% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_classification.R
\name{assign_channels_m5}
\alias{assign_channels_m5}
\title{Assign each indel a channel}
\usage{
assign_channels_m5(indel.df)
}
\arguments{
\item{indel.df}{A classified indel list generated by function indel_classifier89()}
}
\value{
An indel list with channel information
}
\description{
Assign each indel a channel
}
