% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_classification_full.R
\name{assign_channels_mf}
\alias{assign_channels_mf}
\title{Assign each indel a channel from 464 channels}
\usage{
assign_channels_mf(indel.df)
}
\arguments{
\item{indel.df}{A classified indel list generated by function indel_classifier_full()}
}
\value{
An indel list with channel information
}
\description{
Assign each indel a channel from 464 channels
}
