% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_similarity.R
\name{cal_cossim_2sigsets}
\alias{cal_cossim_2sigsets}
\title{Calculate cosine similarity between two signature/profile sets}
\usage{
cal_cossim_2sigsets(sigset1, sigset2, h, w, text_size, ifplot = T, outputname)
}
\arguments{
\item{sigset1}{The first signature set}

\item{sigset2}{The second signature set}

\item{h}{Hight of the plot}

\item{w}{Width of the plot}

\item{text_size}{Size of text}

\item{ifplot}{Plot the output or not, default is True}

\item{outputname}{Output file name of the plot}
}
\value{
A text file including all the cosine similarities between signatures from two data sets
}
\description{
Calculate cosine similarity between two signature/profile sets
}
