% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_similarity.R
\name{cos_similarity}
\alias{cos_similarity}
\title{Calculate cosine similarity between two vectors}
\usage{
cos_similarity(v1, v2)
}
\arguments{
\item{v1}{The first vector}

\item{v2}{The second vector}
}
\value{
The cosine similarity between two vectors
}
\description{
Calculate cosine similarity between two vectors
}
