% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_indel.R
\name{divid_indelcat}
\alias{divid_indelcat}
\title{Separate indel catlogue according to a indel number cutoff}
\usage{
divid_indelcat(mut_cat, cutoff = 500, outputname)
}
\arguments{
\item{mut_cat}{An indel catalogue of multiple samples}

\item{cutoff}{Threshold of indel number to separate samples according to indel burden}

\item{outputname}{file names}
}
\value{
An list of two data.frame
}
\description{
Separate indel catlogue according to a indel number cutoff
}
