% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_file_scott.R
\name{fix_load_indels}
\alias{fix_load_indels}
\title{Fix load indels}
\usage{
fix_load_indels(name)
}
\arguments{
\item{name}{TO BE ADDED}
}
\value{
function TO BE ADDED
}
\description{
Fix load indels
}
