% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_file_scott.R
\name{fix_prepare_indels}
\alias{fix_prepare_indels}
\title{Fix prepare indels}
\usage{
fix_prepare_indels(name)
}
\arguments{
\item{name}{TO BE ADDED}
}
\value{
function TO BE ADDED
}
\description{
Fix prepare indels
}
