% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_indelcatalogue.R
\name{gen_catalogue21}
\alias{gen_catalogue21}
\title{Generate indel catalogue in 21 channels}
\usage{
gen_catalogue21(muts_list, sample_col)
}
\arguments{
\item{muts_list}{A indel list}

\item{sample_col}{Sample column name}
}
\value{
A 21 channel indel catalogue
}
\description{
Generate indel catalogue in 21 channels
}
