% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_indelcatalogue.R
\name{gen_catalogue89}
\alias{gen_catalogue89}
\title{Generate indel catalogue in 89 channels}
\usage{
gen_catalogue89(muts_list, sample_col)
}
\arguments{
\item{muts_list}{A indel list}

\item{sample_col}{Sample column name}
}
\value{
A 89 channel indel catalogue
}
\description{
Generate indel catalogue in 89 channels
}
