% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_indelcatalogue.R
\name{gen_fullcatalogue}
\alias{gen_fullcatalogue}
\title{Generate indel catalogue in full channels}
\usage{
gen_fullcatalogue(muts_list, sample_col)
}
\arguments{
\item{muts_list}{A indel list}

\item{sample_col}{Sample column name}
}
\value{
A full channel indel catalogue
}
\description{
Generate indel catalogue in full channels
}
