% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func21ch.R
\name{gen_plot_catalouge21_single}
\alias{gen_plot_catalouge21_single}
\title{Plot indel profile in a 21-channel bar plot for single sample}
\usage{
gen_plot_catalouge21_single(muts_basis, text_size, plot_title)
}
\arguments{
\item{muts_basis}{A indel catalogue of a single sample}

\item{text_size}{Size of text}

\item{plot_title}{Title of the plot}
}
\value{
A 21-channel indel profile plot
}
\description{
Plot indel profile in a 21-channel bar plot for single sample
}
