% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{gen_plot_catalouge89_single_old}
\alias{gen_plot_catalouge89_single_old}
\title{Old function to plot indel profile/signature in a 89-channel bar plot for single sample}
\usage{
gen_plot_catalouge89_single_old(muts_basis, text_size, plot_title)
}
\arguments{
\item{muts_basis}{A indel catalogue of a single sample}

\item{text_size}{Size of text}

\item{plot_title}{Title of the plot}
}
\value{
A 89-channel indel profile plot
}
\description{
Old function to plot indel profile/signature in a 89-channel bar plot for single sample
}
