% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func_full.R
\name{gen_plot_catalougefull_single_noXlabel}
\alias{gen_plot_catalougefull_single_noXlabel}
\title{Plot indel profile in a extended channel bar plot for single sample
Not show X labels}
\usage{
gen_plot_catalougefull_single_noXlabel(muts_basis, text_size, plot_title)
}
\arguments{
\item{muts_basis}{A indel catalogue of a single sample}

\item{text_size}{Size of text}

\item{plot_title}{Title of the plot}
}
\value{
A extended channel indel profile plot
}
\description{
Plot indel profile in a extended channel bar plot for single sample
Not show X labels
}
