% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_indel.R
\name{hclust_indel}
\alias{hclust_indel}
\title{Hierarchical Clustering to identify high-burden clusters}
\usage{
hclust_indel(
  mut_cat,
  hclust_method = "complete",
  tissue_type,
  ncluster = 6,
  cutoff_height = NULL
)
}
\arguments{
\item{mut_cat}{An indel catalogue of multiple samples}

\item{hclust_method}{This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{tissue_type}{Tissue type}

\item{ncluster}{An integer scalar or vector with the desired number of clusters}

\item{cutoff_height}{Numeric scalar or vector with heights where the tree should be cut}
}
\value{
An object of class hclust which describes the tree produced by the clustering process
}
\description{
Hierarchical Clustering to identify high-burden clusters
}
