% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_classification.R
\name{indel_classifier89}
\alias{indel_classifier89}
\title{Classify indels into different types}
\usage{
indel_classifier89(indels, genome.v)
}
\arguments{
\item{indels}{A indel list. Columns include: "Sample","chr", "position", "REF", "ALT"}

\item{genome.v}{: "hg19", "hg38"}
}
\value{
Classified indel list
}
\description{
Classify indels into different types
}
