% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_classification_full.R
\name{indel_classifier_full}
\alias{indel_classifier_full}
\title{expand 89 channels}
\usage{
indel_classifier_full(indels, genome.v)
}
\arguments{
\item{indels}{A indel list. Columns include: "Sample","chr", "position", "REF", "ALT"}

\item{genome.v}{: "hg19", "hg38"}
}
\value{
Classified indel list
}
\description{
expand 89 channels
}
