% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_highspecific.R
\name{indel_highspecific}
\alias{indel_highspecific}
\title{For signature analysis, indels have to be as clean as possible.
This function will: 1) remove single indels that have original_reps >=10
2) remove nMer indels that have original_reps >=10
3) small indels: length <=100}
\usage{
indel_highspecific(indel.classified)
}
\arguments{
\item{indel.classified}{A classified indel list.}
}
\value{
A filtered indel list
}
\description{
For signature analysis, indels have to be as clean as possible.
This function will: 1) remove single indels that have original_reps >=10
2) remove nMer indels that have original_reps >=10
3) small indels: length <=100
}
