% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_profile_template.R
\docType{data}
\name{indel_template_type_4}
\alias{indel_template_type_4}
\title{indel_template_type_4}
\format{
A data frame with two variables: \code{IndelType}, \code{Indel}.
}
\usage{
indel_template_type_4
}
\description{
Template for 89 channel indel profile
}
\keyword{datasets}
