% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_profile_template.R
\docType{data}
\name{indel_template_type_4_full_figurelabel}
\alias{indel_template_type_4_full_figurelabel}
\title{indel_template_type_4_full_figurelabel}
\format{
A data frame with three variables:
\describe{
\item{IndelType}{Indel type in extended channels}
\item{Indel}{Indel type in 8 channels}
\item{Indel3}{Indel type in 3 channels}
\item{Figlabel}{Lable for Figure}
}
}
\usage{
indel_template_type_4_full_figurelabel
}
\description{
Template for extended channel indel profile
}
\keyword{datasets}
