% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sample.R
\name{normalize_euclidean}
\alias{normalize_euclidean}
\title{Normalize a vector by euclidean length}
\usage{
normalize_euclidean(v)
}
\arguments{
\item{v}{The target vector}
}
\value{
normalized vector
}
\description{
Normalize a vector by euclidean length
}
