% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sample.R
\name{normalize_manhattan}
\alias{normalize_manhattan}
\title{Normalize a vector by manhattan length}
\usage{
normalize_manhattan(v)
}
\arguments{
\item{v}{The target vector}
}
\value{
normalized vector
}
\description{
Normalize a vector by manhattan length
}
