% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sig_distribution.R
\name{plot_sig_count}
\alias{plot_sig_count}
\title{Plot Count of signature in each sample for a given cohort}
\usage{
plot_sig_count(sample_sig_count, SampleCol, h, w, text_size, outputname)
}
\arguments{
\item{sample_sig_count}{A data frame shows the number of mutations of signatures in each sample}

\item{SampleCol}{Sample column name}

\item{h}{Hight of the plot}

\item{w}{Width of the plot}

\item{text_size}{Size of text}

\item{outputname}{Output file name of the plot}
}
\value{
A bar plot of signature distribution (Count) in each sample
}
\description{
Plot Count of signature in each sample for a given cohort
}
