% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func21ch.R
\name{plots_indelprofile_21ch}
\alias{plots_indelprofile_21ch}
\title{Plot indel profile in a 21-channel bar plot, original plots_type_4_m4_89 function}
\usage{
plots_indelprofile_21ch(muts_basis, colnum, h, w, text_size, outputname)
}
\arguments{
\item{muts_basis}{A indel catalogue of multiple samples}

\item{colnum}{Number of columns}

\item{h}{Hight of the plot}

\item{w}{Width of the plot}

\item{text_size}{Size of text}

\item{outputname}{Output file name of the plot}
}
\value{
A plot including 21-channel indel profile of multiple samples
}
\description{
Plot indel profile in a 21-channel bar plot, original plots_type_4_m4_89 function
}
