% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{plots_indelprofile_89ch}
\alias{plots_indelprofile_89ch}
\title{Plot indel profile in a 89-channel bar plot, original plots_type_4_m4_89 function}
\usage{
plots_indelprofile_89ch(
  muts_basis,
  colnum,
  text_size = 3,
  w = NULL,
  h = NULL,
  print_Xlabel = T,
  print_pdf = T,
  print_png = F,
  outputname
)
}
\arguments{
\item{muts_basis}{A indel catalogue of multiple samples}

\item{colnum}{Number of columns}

\item{text_size}{Size of text}

\item{w}{Width of the plot}

\item{h}{Hight of the plot}

\item{print_Xlabel}{print X axis label or not}

\item{print_pdf}{Output pdf file}

\item{print_png}{Output png file}

\item{outputname}{Output file name of the plot}
}
\value{
A plot including 89-channel indel profile of multiple samples
}
\description{
Plot indel profile in a 89-channel bar plot, original plots_type_4_m4_89 function
}
