% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func_full.R
\name{plots_indelprofile_full}
\alias{plots_indelprofile_full}
\title{Plot indel profile in a extended channel bar plot, original plots_type_4_m4_89 function}
\usage{
plots_indelprofile_full(
  muts_basis,
  colnum,
  h,
  w,
  text_size,
  print_Xlabel = T,
  outputname
)
}
\arguments{
\item{muts_basis}{A indel catalogue of multiple samples}

\item{colnum}{Number of columns}

\item{h}{Hight of the plot}

\item{w}{Width of the plot}

\item{text_size}{Size of text}

\item{print_Xlabel}{print X axis label or not}

\item{outputname}{Output file name of the plot}
}
\value{
A plot including extended channel indel profile of multiple samples
}
\description{
Plot indel profile in a extended channel bar plot, original plots_type_4_m4_89 function
}
