% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_file_scott.R
\name{prepare_indels}
\alias{prepare_indels}
\title{Prepare indels sequence context information for classification}
\usage{
prepare_indels(indels, sampleID, genome.v)
}
\arguments{
\item{indels}{A indel list. Columns include: "Sample","chr", "position", "REF", "ALT"}

\item{sampleID}{Cohort name}

\item{genome.v}{: "hg19", "hg38"}
}
\value{
Extended indel list with 5' and 3' sequence information
}
\description{
Prepare indels sequence context information for classification
}
