% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sample.R
\name{preprocess.sample}
\alias{preprocess.sample}
\title{Preprocess samples before signature extraction}
\usage{
preprocess.sample(mut_cat, ss_threshold = 500, rescale.type = "none")
}
\arguments{
\item{mut_cat}{An indel catalogue of multiple samples}

\item{ss_threshold}{Threshold of indel number (default: 500)}

\item{rescale.type}{types of rescale: "manhattan_hb", "euclidean". "manhattan_hb" is to normalize mutation burden of hypermutators to between 2,000-5,000. "euclidean" is 2 norm distance.}
}
\value{
An indel catalogue of multiple samples
}
\description{
Preprocess samples before signature extraction
}
