% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sample.R
\name{rescale.sample}
\alias{rescale.sample}
\title{Plot indel profile in a 89-channel bar plot, original plots_type_4_m4_89 function}
\usage{
rescale.sample(mut_cat, rescale.type)
}
\arguments{
\item{mut_cat}{An indel catalogue of multiple samples}

\item{rescale.type}{types of rescale: "hypermutator", "manhattan", "euclidean". "hypermutator" is to normalize mutation burden of hypermutators to between 2,000-5,000. "manhattan" is 1 norm distance. "euclidean" is 2 norm distance.}
}
\value{
Rescaled indel catalogue to total indel number = 1000
}
\description{
Plot indel profile in a 89-channel bar plot, original plots_type_4_m4_89 function
}
