% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{segment}
\alias{segment}
\title{segmentation function}
\usage{
segment(string, context)
}
\arguments{
\item{string}{A string}

\item{context}{Context of a string}
}
\description{
segmentation function
}
