% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_file_scott.R
\name{segment_indels}
\alias{segment_indels}
\title{Segment indels types for classification}
\usage{
segment_indels(df)
}
\arguments{
\item{df}{An extended indel list generated by prepare_indels().}
}
\value{
Extended indel list with 5' and 3' sequence information
}
\description{
Segment indels types for classification
}
