% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sample.R
\name{ss.sample}
\alias{ss.sample}
\title{Combine low indel burden samples to make synthetic sample with indel number
more than a given threshold}
\usage{
ss.sample(mut_cat, ss_threshold = 500)
}
\arguments{
\item{mut_cat}{An indel catalogue of multiple samples}

\item{ss_threshold}{Threshold of indel number (default: 500)}
}
\value{
An indel catalogue of multiple samples
}
\description{
Combine low indel burden samples to make synthetic sample with indel number
more than a given threshold
}
